Edited the create event page
Added the create event form as a component called EventForm.tsx
Renamed the old create event to page_archive.tsx.
